load cached dataset in data/dataset with two seperate functions for dstart and velocity datasets
