+++ data/dataset.py
@@ -1,10 +1,15 @@
+import hashlib
 import itertools
+import json
+import os
+import pickle
 
 import torch
 import pandas as pd
 import fortepyan as ff
+from omegaconf import DictConfig, OmegaConf
 from tqdm import tqdm
-from datasets import Dataset
+from datasets import Dataset, load_dataset
 
 from data.prepare_dataset import process_record
 from data.tokenizer import Tokenizer, VelocityTokenizer
@@ -212,6 +217,99 @@ class BinsToVelocityDataset(TokenizedMidiDataset):
 
         return vocab_src, vocab_tgt
 
+
+def load_cached_dataset(
+    cfg: DictConfig,
+    split: str = "test",
+) -> TokenizedMidiDataset:
+    if cfg.dataset_class == "BinsToVelocityDataset":
+        return load_velocity_dataset(cfg, split)
+    elif cfg.dataset_class == "BinsToDstartDataset":
+        return load_dstart_dataset(cfg, split)
+
+
+def load_dstart_dataset(cfg, split):
+    n_dstart_bins, n_duration_bins, n_velocity_bins = cfg.bins.split(" ")
+    n_dstart_bins, n_duration_bins, n_velocity_bins = int(n_dstart_bins), int(n_duration_bins), int(n_velocity_bins)
+
+    config_hash = hashlib.sha256()
+    config_string = json.dumps(OmegaConf.to_container(cfg)) + split
+    config_hash.update(config_string.encode())
+    config_hash = config_hash.hexdigest()
+    cache_dir = "tmp/datasets"
+    print(f"Preparing dataset: {config_hash}")
+    try:
+        dataset_cache_file = f"{config_hash}.pkl"
+        dataset_cache_path = os.path.join(cache_dir, dataset_cache_file)
+
+        if os.path.exists(dataset_cache_path):
+            file = open(dataset_cache_path, "rb")
+            dataset = pickle.load(file)
+
+        else:
+            file = open(dataset_cache_path, "wb")
+            hf_dataset = load_dataset(cfg.dataset_name, split=split)
+
+            args = [hf_dataset, n_dstart_bins, n_velocity_bins, n_duration_bins, cfg.sequence_size]
+            dataset = BinsToDstartDataset(
+                dataset=hf_dataset,
+                n_dstart_bins=n_dstart_bins,
+                n_velocity_bins=n_velocity_bins,
+                n_duration_bins=n_duration_bins,
+                sequence_len=cfg.sequence_size,
+                n_tgt_dstart_bins=cfg.tgt_bins,
+            )
+            pickle.dump(dataset, file)
+
+        file.close()
+
+    except (EOFError, ConnectionError, UnboundLocalError):
+        file.close()
+        os.remove(path=dataset_cache_path)
+        dataset = load_cached_dataset(cfg, split)
+
+    return dataset
+
+
+def load_velocity_dataset(cfg, split):
+    n_dstart_bins, n_duration_bins, n_velocity_bins = cfg.bins.split(" ")
+    n_dstart_bins, n_duration_bins, n_velocity_bins = int(n_dstart_bins), int(n_duration_bins), int(n_velocity_bins)
+
+    config_hash = hashlib.sha256()
+    config_string = json.dumps(OmegaConf.to_container(cfg)) + split
+    config_hash.update(config_string.encode())
+    config_hash = config_hash.hexdigest()
+    cache_dir = "tmp/datasets"
+    print(f"Preparing dataset: {config_hash}")
+    try:
+        dataset_cache_file = f"{config_hash}.pkl"
+        dataset_cache_path = os.path.join(cache_dir, dataset_cache_file)
+
+        if os.path.exists(dataset_cache_path):
+            file = open(dataset_cache_path, "rb")
+            dataset = pickle.load(file)
+
+        else:
+            file = open(dataset_cache_path, "wb")
+            hf_dataset = load_dataset(cfg.dataset_name, split=split)
+
+            dataset = BinsToVelocityDataset(
+                dataset=hf_dataset,
+                n_dstart_bins=n_dstart_bins,
+                n_velocity_bins=n_velocity_bins,
+                n_duration_bins=n_duration_bins,
+                sequence_len=cfg.sequence_size,
+            )
+            pickle.dump(dataset, file)
+
+        file.close()
+
+    except (EOFError, ConnectionError, UnboundLocalError):
+        file.close()
+        os.remove(path=dataset_cache_path)
+        dataset = load_cached_dataset(cfg, split)
+
+    return dataset
 
 def main():
     from datasets import load_dataset
